# milktea/skills/milktea-openapi.yaml
openapi: 3.1.0
info:
  title: MilkTea HTTP API — thiết kế một quán
  version: 2.0.0
  description: 'Hợp đồng thiết kế để triển khai; không xác nhận endpoint đã có code. Domain: domain_system.md; HTTP
    envelope: system_architecture.md; security: author_rule.md. Không mô tả STOMP trong OpenAPI. QR read-only; session
    mở khi order/CASHIER open; thanh toán linh hoạt; chỉ CASHIER close. x-allowed-account-roles chỉ ràng buộc nhánh
    xác thực tài khoản, KHÔNG phải điều kiện role cho guest. Guest được cho phép theo nhánh qrAccess/tableAccess/counterAccess
    và phạm vi resource tương ứng; luôn kiểm Domain. Tooling OpenAPI không tự enforce role hay ownership.

    Nguồn: báo cáo MilkTea cập nhật 05/10/2026, 41 UC/T01–T60. Hợp đồng thiết kế; chưa xác nhận các endpoint đã
    được triển khai. v2 bổ sung UC23 thông tin quán, UC30 ngưỡng theo vị trí, UC39 sự cố/tiêu hao pha bù; version
    BIGINT truyền chuỗi; giữ giỏ/context trong kết quả đặt bàn.'
servers:
- url: http://localhost:8080
  description: Ví dụ local; dùng SERVER_PORT và origin thực của project.
paths:
...
  /api/v1/kitchen/orders:
    get:
      operationId: listKitchenOrders
      summary: Bếp xem đơn đã xác nhận kể cả chưa thu
      description: Bếp xem đơn đã xác nhận kể cả chưa thu
      tags:
      - Kitchen
      security:
      - accountCookie: []
      - bearerAuth: []
      responses:
        '200':
          description: Thành công; kiểm tra quyền và invariant phía server.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    const: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/OrderView'
                  meta:
                    $ref: '#/components/schemas/PageMeta'
                additionalProperties: false
                required:
                - success
                - data
                - meta
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '500':
          $ref: '#/components/responses/Error500'
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
        description: ''
      - name: size
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          maximum: 100
          default: 20
        description: ''
      - name: sort
        in: query
        required: false
        schema:
          type: string
          maxLength: 255
        description: Allowlist sort field; default createdAt,desc.
      - name: status
        in: query
        required: false
        schema:
          $ref: '#/components/schemas/OrderStatus'
        description: ''
      x-allowed-account-roles:
      - KITCHEN
      x-use-cases:
      - UC10
      - UC13
...
  /api/v1/kitchen/orders/{orderId}/incidents:
    post:
      operationId: recordOrderIncident
      summary: Báo cáo sự cố đơn hàng
      description: 'Chỉ KITCHEN ghi khi DANG_CHE_BIEN/HOAN_THANH, lý do bắt buộc, audit có actor/time. Kiểm movement
        thuộc đúng đơn nếu liên kết. Không thay OrderStatus, tự hoàn, credit kho hoặc trừ kho; phần chưa ghi gửi
        waste riêng. IncidentView là projection nhật ký: id=MaNK sự kiện ORDER_INCIDENT_RECORDED; payload JSON DuLieu
        chứa movementIds, không bắt buộc bảng Incident riêng.'
      tags:
      - Kitchen
      security:
      - accountCookie: []
      - bearerAuth: []
      x-allowed-account-roles:
      - KITCHEN
      x-use-cases:
      - UC39
      parameters:
      - name: orderId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/Id'
      - name: X-CSRF-TOKEN
        in: header
        required: true
        description: Bắt buộc với browser mutation kể cả guest.
        schema:
          type: string
          maxLength: 1024
      - name: Idempotency-Key
        in: header
        required: true
        description: Key khó đoán; cùng scope/payload trả cùng kết quả, khác payload trả 409.
        schema:
          type: string
          minLength: 16
          maxLength: 64
      responses:
        '201':
          description: Thành công; backend kiểm quyền, trạng thái và transaction.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  success:
                    const: true
                  data:
                    $ref: '#/components/schemas/IncidentView'
                  meta:
                    type: 'null'
                required:
                - success
                - data
                - meta
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '500':
          $ref: '#/components/responses/Error500'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IncidentRequest'
    get:
      operationId: listOrderIncidents
      summary: Xem sự cố đơn hàng
      description: Staff đúng quyền đọc thông tin sự cố; không cấp quyền hoàn thành cho CASHIER/ADMIN.
      tags:
      - Kitchen
      security:
      - accountCookie: []
      - bearerAuth: []
      x-allowed-account-roles:
      - KITCHEN
      - CASHIER
      - ADMIN
      x-use-cases:
      - UC39
      parameters:
      - name: orderId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/Id'
      responses:
        '200':
          description: Thành công; backend kiểm quyền, trạng thái và transaction.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  success:
                    const: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/IncidentView'
                  meta:
                    type: 'null'
                required:
                - success
                - data
                - meta
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '500':
          $ref: '#/components/responses/Error500'
  /api/v1/kitchen/orders/{orderId}/additional-consumptions:
    post:
      operationId: recordAdditionalConsumption
      summary: Ghi tiêu hao pha bù sau hoàn thành
      description: KITCHEN, đơn HOAN_THANH và Incident cùng đơn; chỉ ghi lượng pha bù thực tế tại BEP. Khóa stock,
        không âm, StockMovement TIEU_HAO sourceType=REMAKE/sourceId=chứng từ bổ sung. Không chạy complete lại, không
        đổi hóa đơn/Payment/OrderStatus. Cùng transaction chứng từ+stock+ledger+dedup. Chứng từ pha bù là audit
        ORDER_REMAKE_CONSUMED có MaNKCha=incidentId; sourceId=MaNK chứng từ, sourceLine=materialId. Audit và ledger
        cùng transaction; lưu items để truy vết, không dùng số ID sự cố làm ID của mọi lần pha bù.
      tags:
      - Kitchen
      security:
      - accountCookie: []
      - bearerAuth: []
      x-allowed-account-roles:
      - KITCHEN
      x-use-cases:
      - UC39
      - UC34
      parameters:
      - name: orderId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/Id'
      - name: X-CSRF-TOKEN
        in: header
        required: true
        description: Bắt buộc với browser mutation kể cả guest.
        schema:
          type: string
          maxLength: 1024
      - name: Idempotency-Key
        in: header
        required: true
        description: Key khó đoán; cùng scope/payload trả cùng kết quả, khác payload trả 409.
        schema:
          type: string
          minLength: 16
          maxLength: 64
      responses:
        '201':
          description: Thành công; backend kiểm quyền, trạng thái và transaction.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  success:
                    const: true
                  data:
                    $ref: '#/components/schemas/InventoryDocumentView'
                  meta:
                    type: 'null'
                required:
                - success
                - data
                - meta
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '500':
          $ref: '#/components/responses/Error500'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdditionalConsumptionRequest'
...
  /api/v1/kitchen/orders/{orderId}/incidents/{incidentId}/resolve:
    post:
      operationId: resolveOrderIncident
      summary: Ghi nhận kết quả xử lý sự cố đơn hàng
      description: KITCHEN xác nhận xử lý Incident đúng đơn, ghi lý do và audit append-only. Không tự complete,
        ghi hao hụt, pha bù, sửa tiền hay kho. Khóa Incident gốc, từ OPEN sang RESOLVED đúng một lần; retry cùng
        idempotency trả kết quả cũ. Đây là hoàn thiện kỹ thuật UC39; không thêm điều kiện chặn đóng phiên ngoài
        nghiệp vụ đã chốt.
      tags:
      - Kitchen
      security:
      - accountCookie: []
      - bearerAuth: []
      x-allowed-account-roles:
      - KITCHEN
      x-use-cases:
      - UC39
      parameters:
      - name: orderId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/Id'
      - name: incidentId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/Id'
      - name: X-CSRF-TOKEN
        in: header
        required: true
        description: Bắt buộc với browser mutation kể cả guest.
        schema:
          type: string
          maxLength: 1024
      - name: Idempotency-Key
        in: header
        required: true
        description: Key khó đoán; cùng scope/payload trả cùng kết quả, khác payload trả 409.
        schema:
          type: string
          minLength: 16
          maxLength: 64
      responses:
        '201':
          description: Thành công; backend kiểm quyền, trạng thái và transaction.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  success:
                    const: true
                  data:
                    $ref: '#/components/schemas/IncidentView'
                  meta:
                    type: 'null'
                required:
                - success
                - data
                - meta
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '500':
          $ref: '#/components/responses/Error500'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReasonRequest'
components:
  securitySchemes:
    accountCookie:
      type: apiKey
      in: cookie
      name: milktea_access_token
      description: JWT account HttpOnly; role/state/version enforced server-side.
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    tableAccess:
      type: apiKey
      in: header
      name: X-Table-Token
      description: Opaque access scoped to one currently OPEN TableSession, not an account role.
    counterAccess:
      type: apiKey
      in: header
      name: X-Counter-Token
      description: Opaque access scoped to one counter order; client keeps only in current page RAM.
    qrAccess:
      type: apiKey
      in: header
      name: X-QR-Code
      description: Opaque QR code identifies a table; never a staff credential.
  schemas:
    Id:
      type: string
      pattern: ^[1-9][0-9]*$
      example: '5'
    Decimal:
      type: string
      pattern: ^[0-9]+(\.[0-9]{1,3})?$
      example: '35000'
...
    OrderStatus:
      type: string
      enum:
      - CHO_XAC_NHAN
      - CHO_CHE_BIEN
      - DANG_CHE_BIEN
      - HOAN_THANH
      - DA_HUY
    OrderSource:
      type: string
      enum:
      - TABLE
      - COUNTER
...
    PaymentMethod:
      type: string
      enum:
      - CASH
      - BANK_TRANSFER
    ReceiptState:
      type: string
      enum:
      - CHUA_THU
      - DA_THU
    InvoiceStatus:
      type: string
      enum:
      - HIEU_LUC
      - DA_HUY
    CancellationStatus:
      type: string
      enum:
      - PENDING
      - APPROVED
      - REJECTED
      description: API PENDING/APPROVED/REJECTED ánh xạ DB CHO/CHAP_THUAN/TU_CHOI bằng converter rõ ràng; không
        dùng EnumType.ORDINAL hoặc lưu enum.name() khác hợp đồng DB.
    RefundStatus:
      type: string
      enum:
      - CHO_DUYET
      - DA_DUYET
      - TU_CHOI
      - DA_HOAN
...
    PageMeta:
      type: object
      properties:
        page:
          type: integer
          minimum: 0
        size:
          type: integer
          minimum: 0
        totalElements:
          type: integer
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
      additionalProperties: false
      required:
      - page
      - size
      - totalElements
      - totalPages
    ErrorResponse:
      type: object
      properties:
        success:
          const: false
        error:
          type: object
          properties:
            code:
              type: string
              maxLength: 80
            message:
              type: string
              maxLength: 1000
            fieldErrors:
              type: object
              additionalProperties:
                type: string
                maxLength: 1000
          additionalProperties: false
          required:
          - code
          - message
          - fieldErrors
        traceId:
          type: string
          maxLength: 100
      additionalProperties: false
      required:
      - success
      - error
      - traceId
...
    OrderItemView:
      type: object
      properties:
        productId:
          $ref: '#/components/schemas/Id'
        name:
          type: string
          maxLength: 120
        size:
          type: string
          maxLength: 30
        quantity:
          type: integer
          minimum: 1
          maximum: 100
        unitPrice:
          $ref: '#/components/schemas/Money'
        lineTotal:
          $ref: '#/components/schemas/Money'
      additionalProperties: false
      required:
      - productId
      - name
      - size
      - quantity
      - unitPrice
      - lineTotal
    PaymentView:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/Id'
        invoiceId:
          $ref: '#/components/schemas/Id'
        amount:
          $ref: '#/components/schemas/Money'
        method:
          $ref: '#/components/schemas/PaymentMethod'
        paidAt:
          type: string
          format: date-time
        reference:
          anyOf:
          - type: string
            maxLength: 255
          - type: 'null'
      additionalProperties: false
      required:
      - id
      - invoiceId
      - amount
      - method
      - paidAt
      - reference
    InvoiceView:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/Id'
        status:
          $ref: '#/components/schemas/InvoiceStatus'
        subtotal:
          $ref: '#/components/schemas/Money'
        discountPercent:
          $ref: '#/components/schemas/Decimal'
        discountAmount:
          $ref: '#/components/schemas/Money'
        total:
          $ref: '#/components/schemas/Money'
        receiptState:
          $ref: '#/components/schemas/ReceiptState'
        outstandingAmount:
          $ref: '#/components/schemas/Money'
      additionalProperties: false
      required:
      - id
      - status
      - subtotal
      - discountPercent
      - discountAmount
      - total
      - receiptState
      - outstandingAmount
    CancellationView:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/Id'
        orderId:
          $ref: '#/components/schemas/Id'
        status:
          $ref: '#/components/schemas/CancellationStatus'
        reason:
          type: string
          maxLength: 500
        decisionReason:
          anyOf:
          - type: string
            maxLength: 500
          - type: 'null'
      additionalProperties: false
      required:
      - id
      - orderId
      - status
      - reason
      - decisionReason
    RefundView:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/Id'
        orderId:
          $ref: '#/components/schemas/Id'
        paymentId:
          $ref: '#/components/schemas/Id'
        amount:
          $ref: '#/components/schemas/Money'
        status:
          $ref: '#/components/schemas/RefundStatus'
        refundedAt:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
        reference:
          anyOf:
          - type: string
            maxLength: 255
          - type: 'null'
      additionalProperties: false
      required:
      - id
      - orderId
      - paymentId
      - amount
      - status
      - refundedAt
      - reference
    OrderView:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/Id'
        source:
          $ref: '#/components/schemas/OrderSource'
        sessionId:
          anyOf:
          - $ref: '#/components/schemas/Id'
          - type: 'null'
        status:
          $ref: '#/components/schemas/OrderStatus'
        createdAt:
          type: string
          format: date-time
        completedAt:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItemView'
        invoice:
          $ref: '#/components/schemas/InvoiceView'
        payments:
          type: array
          items:
            $ref: '#/components/schemas/PaymentView'
        cancellations:
          type: array
          items:
            $ref: '#/components/schemas/CancellationView'
        refunds:
          type: array
          items:
            $ref: '#/components/schemas/RefundView'
        replacementOfOrderId:
          anyOf:
          - $ref: '#/components/schemas/Id'
          - type: 'null'
      additionalProperties: false
      required:
      - id
      - source
      - sessionId
      - status
      - createdAt
      - completedAt
      - items
      - invoice
      - payments
      - cancellations
      - refunds
      - replacementOfOrderId
...
    ReasonRequest:
      type: object
      properties:
        reason:
          type: string
          minLength: 1
          maxLength: 500
      additionalProperties: false
      required:
      - reason
...
    IngredientInput:
      type: object
      properties:
        materialId:
          $ref: '#/components/schemas/Id'
        quantity:
          $ref: '#/components/schemas/Decimal'
      additionalProperties: false
      required:
      - materialId
      - quantity
...
    InventoryDocumentView:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/Id'
        type:
          type: string
          enum:
          - IMPORT
          - ISSUE
          - PREPARATION
          - WASTE
          - ADDITIONAL_CONSUMPTION
        createdAt:
          type: string
          format: date-time
        items:
          type: array
          items:
            $ref: '#/components/schemas/InventoryLineView'
        reason:
          type: string
          maxLength: 500
      additionalProperties: false
      required:
      - id
      - type
      - createdAt
      - items
      - reason
      description: unitCost có giá trị trên phiếu nhập theo ChiTietNhap, null cho xuất/hao hụt/tiêu hao bổ sung;
        bản chốt giá nhập không đổi khi sửa dữ liệu nền.
...
    Money:
      type: string
      pattern: ^[0-9]+(\.[0-9]{1,2})?$
      description: Tiền VND không âm; scale tối đa 2, backend BigDecimal; làm tròn HALF_UP tới đồng khi tính hóa
        đơn.
      example: '63000'
...
    IncidentRequest:
      type: object
      additionalProperties: false
      properties:
        reason:
          type: string
          minLength: 1
          maxLength: 1000
        relatedMovementIds:
          type: array
          maxItems: 100
          items:
            $ref: '#/components/schemas/Id'
      required:
      - reason
    IncidentView:
      type: object
      additionalProperties: false
      properties:
        id:
          $ref: '#/components/schemas/Id'
        orderId:
          $ref: '#/components/schemas/Id'
        reason:
          type: string
        createdAt:
          type: string
          format: date-time
        relatedMovementIds:
          type: array
          items:
            $ref: '#/components/schemas/Id'
        status:
          type: string
          enum:
          - OPEN
          - RESOLVED
        resolvedAt:
          type:
          - string
          - 'null'
          format: date-time
        resolutionReason:
          type:
          - string
          - 'null'
      required:
      - id
      - orderId
      - reason
      - createdAt
      - relatedMovementIds
      - status
      - resolvedAt
      - resolutionReason
    AdditionalConsumptionRequest:
      type: object
      additionalProperties: false
      properties:
        incidentId:
          $ref: '#/components/schemas/Id'
        items:
          type: array
          minItems: 1
          maxItems: 100
          items:
            $ref: '#/components/schemas/IngredientInput'
        reason:
          type: string
          minLength: 1
          maxLength: 500
      required:
      - incidentId
      - items
      - reason
...
    InventoryLineView:
      type: object
      additionalProperties: false
      properties:
        materialId:
          $ref: '#/components/schemas/Id'
        quantity:
          $ref: '#/components/schemas/Decimal'
        unitCost:
          anyOf:
          - $ref: '#/components/schemas/Money'
          - type: 'null'
      required:
      - materialId
      - quantity
      - unitCost
  responses:
    Error400:
      description: HTTP 400; error code/shape theo system_architecture.md.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Error401:
      description: HTTP 401; error code/shape theo system_architecture.md.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Error403:
      description: HTTP 403; error code/shape theo system_architecture.md.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Error404:
      description: HTTP 404; error code/shape theo system_architecture.md.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Error409:
      description: HTTP 409; error code/shape theo system_architecture.md.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    Error500:
      description: HTTP 500; error code/shape theo system_architecture.md.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
security:
- accountCookie: []
- bearerAuth: []
...